+++ packages/evolution-backend/src/api/survey.participant.routes.ts
@@ -79,7 +79,7 @@ export default (authorizationMiddleware, loggingMiddleware: InterviewLoggingMidd
                         (req.user as UserAttributes).id,
                         {},
                         loggingMiddleware.getUserIdForLogging(req),
-                        ['id', 'uuid', 'is_valid', 'is_completed', 'response', 'participant_id']
+                        ['id', 'uuid', 'is_completed', 'response', 'participant_id']
                     );
                 }